requirements.txt
pandas
pypdf
python-dotenv
langchain>=0.1.20
langchain-core>=0.1.33
langchain-community>=0.0.30
langchainhub>=0.1.14
langchain-groq>=0.0.2
sentence-transformers>=2.6.1
faiss-cpu>=1.7.4
groq>=0.5.0
pyodbc>=5.0.1
sqlalchemy>=2.0.30
langgraph
streamlit
